% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{ageutils-defunct}
\alias{ageutils-defunct}
\alias{reaggregate_interval_rates}
\alias{reaggregate_interval_counts}
\alias{split_interval_counts}
\alias{aggregate_age_counts}
\title{Defunct functions}
\usage{
reaggregate_interval_rates(
  lower_bounds,
  upper_bounds = NULL,
  rates,
  breaks,
  weights = NULL
)

reaggregate_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  breaks,
  max_upper = 100L,
  weights = NULL
)

split_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  max_upper = 100L,
  weights = NULL
)

aggregate_age_counts(counts, ages = seq_along(counts) - 1L, breaks)
}
\description{
Functions removed from ageutils in the 0.0.8 release. They have been replaced
with stubs that will error and report them as defunct.
\itemize{
\item \code{reaggregate_interval_rates()} (best replaced by \code{reaggregate_rates()}),
\item \code{reaggregate_interval_counts()} (best replaced by \code{reaggregate_counts()}),
\item \code{split_interval_counts()}, and
\item \code{aggregate_age_counts()}.
}
}
\keyword{internal}
