% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ageutils-deprecated}
\alias{ageutils-deprecated}
\alias{reaggregate_interval_rates}
\alias{reaggregate_interval_counts}
\alias{split_interval_counts}
\alias{aggregate_age_counts}
\title{Deprecated functions}
\usage{
reaggregate_interval_rates(
  lower_bounds,
  upper_bounds = NULL,
  rates,
  breaks,
  weights = NULL
)

reaggregate_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  breaks,
  max_upper = 100L,
  weights = NULL
)

split_interval_counts(
  lower_bounds,
  upper_bounds,
  counts,
  max_upper = 100L,
  weights = NULL
)

aggregate_age_counts(counts, ages = seq_along(counts) - 1L, breaks)
}
\description{
The functions listed below are deprecated and may be removed in a future
package release.
}
\note{
Help pages for each deprecated function can be found via
\code{help("<function>-deprecated")}.
}
\seealso{
\code{\link[=reaggregate_rates]{reaggregate_rates()}}:  A replacement for \code{reaggregate_interval_rates()}.

\code{\link[=reaggregate_counts]{reaggregate_counts()}}:  A replacement for \code{reaggregate_interval_counts()}.
}
\keyword{internal}
