% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageutils-package.R
\docType{package}
\name{ageutils-package}
\alias{ageutils}
\alias{ageutils-package}
\title{ageutils: Collection of Functions for Working with Age Intervals}
\description{
Provides a collection of efficient functions for working with individual ages and corresponding intervals. These include functions for conversion from an age to an interval, aggregation of ages with associated counts in to intervals and the splitting of interval counts based on specified age distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://timtaylor.github.io/ageutils/}
  \item Report bugs at \url{https://github.com/TimTaylor/ageutils/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Edwin van Leeuwen \email{edwinvanl@gmail.com} (\href{https://orcid.org/0000-0002-2383-5305}{ORCID}) [contributor]
}

}
\keyword{internal}
