% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_dat}
\alias{pop_dat}
\title{Aggregated population data}
\format{
A data frame with 200 rows and 6 variables:
\describe{
\item{area_code}{Unique area identifier}
\item{area_name}{Unique area name}
\item{age_category}{Left-closed and right-open age interval}
\item{value}{count of individ}
}
}
\source{
\url{https://github.com/TimTaylor/census_pop_2021}
}
\usage{
pop_dat
}
\description{
A dataset derived from the 2021 UK census containing population for different
age categories across England and Wales.
}
\keyword{datasets}
