% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaggregate_counts.R
\name{reaggregate_counts}
\alias{reaggregate_counts}
\alias{reaggregate_counts.default}
\title{Reaggregate age counts}
\usage{
reaggregate_counts(...)

\method{reaggregate_counts}{default}(
  bounds,
  counts,
  new_bounds,
  ...,
  population_bounds = NULL,
  population_weights = NULL
)
}
\arguments{
\item{...}{Arguments passed to underlying methods}

\item{bounds}{\verb{[numeric]}}

\item{counts}{\verb{[numeric]}}

\item{new_bounds}{\verb{[numeric]}}

\item{population_bounds}{\verb{[numeric]}}

\item{population_weights}{\verb{[numeric]}}
}
\description{
Reaggregate age counts
}
