% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaggregate_rates.R
\name{reaggregate_rates}
\alias{reaggregate_rates}
\alias{reaggregate_rates.default}
\title{Reaggregate age rates}
\usage{
reaggregate_rates(...)

\method{reaggregate_rates}{default}(
  bounds,
  rates,
  new_bounds,
  ...,
  population_bounds = NULL,
  population_weights = NULL
)
}
\arguments{
\item{...}{Arguments passed to underlying methods}

\item{bounds}{\verb{[numeric]}}

\item{rates}{\verb{[numeric]}}

\item{new_bounds}{\verb{[numeric]}}

\item{population_bounds}{\verb{[numeric]}}

\item{population_weights}{\verb{[numeric]}}
}
\description{
Reaggregate age rates
}
\examples{
reaggregate_rates(
    bounds = c(0, 5, 10),
    rates = c(0.1, 0.2 ,0.3),
    new_bounds = c(0, 2, 7, 10),
    population_bounds = c(0, 2, 5, 7, 10),
    population_weights = c(100, 200, 50, 150, 100)
)

}
